BERT
Pre-training: The model is trained on a large corpus of text using unsupervised tasks like masked language modeling (MLM) and next sentence prediction (NSP). In MLM, some words in a sentence are randomly masked, and BERT learns to predict them. In NSP, the model is trained to predict if two sentences follow each other.